requirements.txt
textual
python-dotenv
tomli
httpx
ollama
pyperclip
